requirements.txt
# Core dependencies
langgraph>=0.0.19
langgraph-swarm>=0.0.5
python-dotenv>=1.0.0

# LLM provider dependencies
anthropic>=0.8.0  # For Anthropic Claude
openai>=1.3.0     # For OpenAI GPT models
perplexity-python>=0.1.0  # For Perplexity models

# Utility dependencies
tqdm>=4.66.1
rich>=13.6.0
pydantic>=2.4.2

requests>=2.25.0
colorama>=0.4.4
markdown>=3.3.0
tiktoken>=0.4.0
tenacity>=8.0.0
regex>=2022.1.18 